Added initial pulse calculation and string conversion
When a pulse is detected (currently emulated using a button) the
current power consumption is calculated and displayed on the screen.
The value for “Current (W)” is now actually based on the float value.
Total is based on the float value, but is not updated properly. Today
is not implemented yet, as we don’t have time yet.