Board update, fixed some unfortunate mistakes on my part.
The oak only has a single Analog pin, I attached the sensor to one of
the digital pins. I can bodge the existing boards, but I’ve chosen to
order new boards anyway. So this will be version 2.0, also containing a
Ground pin for the UART connection as well as bigger holes for the IR
sensor and a jumper to set the module in ‘bootloader’ mode.